requirements_app.txt
pytorch-lightning==1.2.10
datasets==2.4.0
scikit-learn==0.24.2
huggingface_hub==0.1.0
hydra-core
omegaconf
hydra_colorlog
onnxruntime
fastapi
uvicorn